elysia deno integration
Honestly speaking it was so smooth!
And even better then WebStatndard official adapter or node one.
Why?
Because node adapter though can be run with Deno made extra wrapping
around WebStandard one.
So Deno is already ready to utilize WebStandard adapter.
But why it even better with Deno?
Because Deno can dead-simply upgrade GET request to WEBSOCKET
(though elysia's ws feature is not supported... it may be not needed
now)