rest main
Explicitly split possible program variants at the early start.
Even before actual config initialization.

So the main idea is to name possible modes of program.

As a result - no matter how difference configurations may be - they will
be simply choosed, switched.

Such variant of program is named "MODE".

However at least this variable should be defined.
Also possible that another variables of such absolute basic stuff
will be extend this config.

But important to understand - this mode should be computed with such
rules:
1. As possibly early
2. All requirements should have default - so some mode, for example
   "local" may be computed without any actions
3. Though technically this is configuration, it should be splited from
   main one. No project related stuff only.

The main purpose is to decide in which mode application should be run.